# Conflicts:
#	engineer_middleware/CMakeLists.txt
#	engineer_middleware/include/engineer_middleware/chassis_interface.h
#	engineer_middleware/include/engineer_middleware/middleware.h
#	engineer_middleware/src/middleware.cpp
#	engineer_middleware/src/step_queue.cpp

+++ engineer_middleware/include/engineer_middleware/motion.h
@@ -134,7 +134,7 @@ class JointMotion : public MoveitMotionBase {
   bool isFinish() override {
     std::vector<double> current = interface_.getCurrentJointValues();
     double error = 0.;
-    for (int i = 0; i < current.size(); ++i)
+    for (int i = 0; i < (int) current.size(); ++i)
       error += std::abs(target_[i] - current[i]);
     return error < tolerance_angular_;
   }
